upload student page now allows for uploading existing students to new sections

+++ public/uploadStudentList.php
@@ -14,39 +14,59 @@ try {
     echo "<br>" . $error->getMessage();
 }
 
+try {
+    $all_students = User::where(array('is_instructor' => '0', 'is_admin' => '0'));
+} catch (PDOException $error) {
+    echo "<br>" . $error->getMessage();
+}
 
     $create_success = false;
-
 
     if (isset($_POST['submit2'])) {
         $students = fopen($_POST["fileID"], "r");
         $headers = fgetcsv($students, 1000, ",");
         $count = 0;
         while (($studentData = fgetcsv($students, 1000, ",")) !== false) {
-            $user = new User();
-            $user->student_id = $studentData[0];
-            $user->first_name = $studentData[1];
-            $user->last_name = $studentData[2];
-            $user->email = $studentData[3];
-            $user->is_admin = 0;
-            $user->is_instructor = 0;
-            $user->is_ta = 0;
-            $user->password_digest = password_hash("welcome", PASSWORD_DEFAULT);
 
-            try {
-                $user->save();
-                $count++;
-                $create_success = true;
-            } catch (PDOException $error) {
-                echo "<br>" . $error->getMessage();
-            }
-            $section_student = new SectionStudent();
-            $section_student->user_id = $user->id;
-            $section_student->section_id = $_POST['sectionID'];
-            try {
-                $section_student->save();
-            } catch (PDOException $error) {
-                echo "<br>" . $error->getMessage();
+            $existing_student = array_search($studentData[0], array_column($all_students, 'student_id'));
+
+            if(is_null($existing_student)){
+                $user = new User();
+                $user->student_id = $studentData[0];
+                $user->first_name = $studentData[1];
+                $user->last_name = $studentData[2];
+                $user->email = $studentData[3];
+                $user->is_admin = 0;
+                $user->is_instructor = 0;
+                $user->is_ta = 0;
+                $user->password_digest = password_hash("welcome", PASSWORD_DEFAULT);
+
+                try {
+                    $user->save();
+                    $count++;
+                    $create_success = true;
+                } catch (PDOException $error) {
+                    echo "<br>" . $error->getMessage();
+                }
+                $section_student = new SectionStudent();
+                $section_student->user_id = $user->id;
+                $section_student->section_id = $_POST['sectionID'];
+                $section_success = true;
+                try {
+                    $section_student->save();
+                } catch (PDOException $error) {
+                    echo "<br>" . $error->getMessage();
+                }
+            } else {
+                $section_student = new SectionStudent();
+                $section_student->user_id = $all_students[$existing_student]->id;
+                $section_student->section_id = $_POST['sectionID'];
+                $section_success = true;
+                try {
+                    $section_student->save();
+                } catch (PDOException $error) {
+                    echo "<br>" . $error->getMessage();
+                }
             }
 
         }
@@ -62,7 +82,7 @@ try {
 <?php
     if (isset($_POST['submit2']) && $create_success) {
         echo "<h3> $count Students added successfully</h3><br>";
-    } elseif (isset($_POST['submit2']) && !$create_success) {
+    } elseif (isset($_POST['submit2']) && (!$create_success Or !$section_success)) {
         echo "<h3> Student List upload failed </h3>";
     }
 ?>
